+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Introduce new `IsRepositoryNotFound` error matcher
+
 ## [0.1.1] - 2020-03-17
 
 ### Added